requirements.txt
pyodbc==4.0.39
pillow==10.2.0
reportlab==4.1.0
pandas==2.2.0
seaborn==0.13.2
matplotlib==3.8.2
pyjwt==2.8.0
cryptography==42.0.2
python-dateutil==2.8.2
autopep8==2.0.4
pylint==3.0.3
PyJWT==2.8.0